Moved all author code, out of library file
